LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStream-default-feature-policy.https.sub-expected.txt

CONSOLE MESSAGE: line 1: Trying to call getUserMedia from a document with a different security origin than its top-level frame.
CONSOLE MESSAGE: line 1: Trying to call getUserMedia from a document with a different security origin than its top-level frame.
CONSOLE MESSAGE: line 1: Trying to call getUserMedia from a document with a different security origin than its top-level frame.
CONSOLE MESSAGE: line 1: Trying to call getUserMedia from a document with a different security origin than its top-level frame.
CONSOLE MESSAGE: line 1: Trying to call getUserMedia from a document with a different security origin than its top-level frame.


PASS Default "microphone" feature policy ["self"] allows the top-level document. 
PASS Default "microphone" feature policy ["self"] allows same-origin iframes. 
FAIL Default "microphone" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#[object NavigatorUserMediaError]" but got "#NotAllowedError: The request is not allowed by the user agent or the platform in the current context, possibly because the user denied permission."
FAIL Feature policy "microphone" can be enabled in cross-origin iframes using "allow" attribute. undefined is not an object (evaluating 'frame.allow.concat')
PASS Default "camera" feature policy ["self"] allows the top-level document. 
PASS Default "camera" feature policy ["self"] allows same-origin iframes. 
FAIL Default "camera" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#[object NavigatorUserMediaError]" but got "#NotAllowedError: The request is not allowed by the user agent or the platform in the current context, possibly because the user denied permission."
FAIL Feature policy "camera" can be enabled in cross-origin iframes using "allow" attribute. undefined is not an object (evaluating 'frame.allow.concat')
PASS Default "camera; microphone" feature policy ["self"] allows the top-level document. 
PASS Default "camera; microphone" feature policy ["self"] allows same-origin iframes. 
FAIL Default "camera; microphone" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#[object NavigatorUserMediaError]" but got "#NotAllowedError: The request is not allowed by the user agent or the platform in the current context, possibly because the user denied permission."
FAIL Feature policy "camera; microphone" can be enabled in cross-origin iframes using "allow" attribute. undefined is not an object (evaluating 'frame.allow.concat')
